requirements.txt
# AgentLedger – Python dependencies
# Core
pydantic>=2.6,<3
pandas>=2.2,<3
langgraph>=0.2,<1
langchain>=0.3,<1
langchain-openai>=0.2,<1
langchain-huggingface>=0.1,<1
langchain-community>=0.3,<1
sentence-transformers>=3.0,<4
faiss-cpu>=1.8,<2
streamlit>=1.30,<2

# Environment
python-dotenv>=1.0,<2

# Mock data generation
faker>=28.0,<30

# Dev / testing
pytest>=8.0,<9
